UI Main Menu and Camera Movement
Added script and prefab UIInteractionController. This allows you to set start and end coordinates for menu items and does animation movements for said items. You can set a movement type and the delay an animation takes to move. Takes the cameracontoller script to be able to call onGameplay so the rest of the camera controller script animates the camera and performs the rest of the code after the user clicks play.

Added Canvas prefab with prefab children for UI.

Edited CameraController script to perform camera zoom in animation.