UI and Camera with Movement
Added codebase for DoTween And TextMeshPro.

Added Canvas and UIMenuControl game objects to scene.

UserInterfaceController - Takes the UI game objects and allows you to edit their position animation movements as well as their type of movement. Calls CameraController to move the camera to the initial player object to focus on.

CameraController - Edited script to have OnGamePlay which sets the characterController and starts the camera movement to the initial player game object. After movement is complete isInGame property gets set to true allowing the rest of the code in update to function. Added 3 fields, the movement type for the camera, the movement delay for the camera and the initial object the camera will go to.

FlickingMechanics - Added private CameraController object that gets set on awake. Added if statement to check the IsInGame Property to see if its true before calling the rest of the code.