UI Main Menu and Camera Movement
Added code base for DoTween and Text Mesh Pro

Added script and prefab UIInteractionController - This allows you to set start and end coordinates for menu items and does animation movements for said items. You can set a movement type and the delay an animation takes to move. Takes the cameracontoller script to be able to call OnGameplay so the rest of the camera controller script animates the camera and performs the rest of the code after the user clicks play.

Added Canvas prefab with prefab children for UI.

CameraController-  Edited script to perform camera zoom in animation. Added fields to control movement type and delay. Added bool property to check when gameplay is ready to start.

Flicking Mechanic- Added cameraController field that gets set on awake and use on Fixed update to check whether or not the game has started before running the rest of the code.