UI Tutorial Slide and Transition Mechanic
Updated UserInteractioncontrol to add to more fields that take the tutorial image and transition objects. Added code to handle moving tutorial image to screen after a set amount of time determined by a field. Added transition function that can be used once called to move transition image from the bottom to the top of the screen.