UI PlayerState, Error Handling and TreeHolder
Game now starts with the state machine in the UI playerState. Once the tutorial screen is hidden, the playerState is changed back to Pinecone_StandingBy to allow for input.

For error handling, if the state machine does not have a reference to a tree object, then player can just keep flicking the pinecone. Also, if there is not already a TreeHolder in the scene, the state machine now creates a Treeholder.