End UI Menu
Added prefab to canvas that is the menu when a player finishes. Updated scene with new prefab.

Edited UserInteractionControl to have scale animation capabilities in case you want to scale things. Added field for the levelFinishMenu. Removed trainsitionEndPoint field as it was not used in the rest of the code.

Added Main Level scene to build settings so that the play again button will reload the scene.