DoTween Transition Fix and Camera Arial View
Userinteraction Control - Added fix for error where dotween was logging max tween counts. This happened because the transition method did not kill the tween from completion.

Added function that sets the main camera to a vector 3 of the users choosing called MoveCameraToArialView. This will allow the camera to cut to a position set when the IsGamePlay is set to false.